We've made some significant improvements to our application to enhance its performance and prepare for exciting future developments. Here's a quick overview of the key changes:
Easier Version Tracking: We've introduced a new way to keep track of the version of our application. This is important because it sets the stage for some cool things we have planned, like automating tasks using GitHub Actions.

Smoother Operation: We've smoothed out some bumps in the road. The CheckForUpdate() and PrepareUpdate() functions in the background now run more smoothly without any awkward pauses.