The most important change is the addition of a new property called `CountOculusSystemAsVR` in the `DataAndSecurity` namespace of the `DataController.cs` file. This property is of type `bool` and is added under the category "System" (DataController.cs).
In the `ComponentStatsModule.cs` file, the `IsVRRunning()` method in the `Classes.Modules` namespace has been modified. A new condition has been added to check if `ViewModel.Instance.CountOculusSystemAsVR` is true. If it is true, the method also checks if the process "OVRServer_x64" is running (ComponentStatsModule.cs).

In the `ViewModel.cs` file, the `ViewModels` namespace has added a new property called `CountOculusSystemAsVR`. This property is of type `bool` and is used to get and set the value of `_CountOculusSystemAsVR`. It also notifies the property changed event when the value is set (ViewModel.cs).

References:
- DataController.cs (namespace `vrcosc_magicchatbox.DataAndSecurity`): Added a new property called `CountOculusSystemAsVR` of type `bool` under the category "System".
- ComponentStatsModule.cs (namespace `vrcosc_magicchatbox.Classes.Modules`): Modified the `IsVRRunning()` method to include a new condition to check if `ViewModel.Instance.CountOculusSystemAsVR` is true and if the process "OVRServer_x64" is running.
- ViewModel.cs (namespace `vrcosc_magicchatbox.ViewModels`): Added a new property called `CountOculusSystemAsVR` of type `bool` to get and set the value of `_CountOculusSystemAsVR` and notify the property changed event.