Updated `_ApiStream` value in `ViewModel.cs`
Updated the value of the `_ApiStream` private string variable in the `ViewModels` namespace of the `ViewModel.cs` file. The previous value was a long string, and it has been changed to a new string. This change is part of an ongoing effort to improve the security and efficiency of our API calls.