Refactor IntelliChat methods and update parameters
This commit includes a series of changes to improve the IntelliChat codebase. Unused namespaces `System`, `System.Text`, and `System.Windows.Threading` were removed. The `LanguageName` property in the `SupportedIntelliChatLanguage` class was renamed to `Language`. The `PerformSpellingAndGrammarCheckAsync` method underwent several changes, including the removal of the `languages` parameter, which was replaced with a check for `_isInitialized`, and the replacement of `ViewModel.Instance.IntelliChatPerformModeration` with `Settings.IntelliChatPerformModeration`. The use of `languages` was replaced with `SelectedSupportedLanguages`. The `PerformBeautifySentenceAsync` method now includes a new `intelliChatWritingStyle` parameter and the use of `SelectedWritingStyle` was replaced with `writingStyle`. A check for `_isInitialized` was added to the `GenerateConversationStarterAsync` method. The `PerformLanguageTranslationAutoDetectAsync` method now uses `supportedIntelliChatLanguage` instead of `text` and `intelliChatLanguage` instead of `targetLanguage`. Lastly, the use of `ViewModel.Instance` was replaced with `Settings` in the `AcceptIntelliChatSuggestion`, `RejectIntelliChatSuggestion`, and `PerformModerationCheckAsync` methods.