Updated UI elements in MainWindow.xaml
Updated various UI elements in the MainWindow.xaml file. Removed the "IntgrScanForc_checkbox" CheckBox and added a new CheckBox "LimitTitleOnsApp_checkbox" for displaying the desktop focused app. Introduced two TextBoxes "WindowActivityDesktopTitle" and "WindowActivityDesktopFocusTitle" for displaying the title of the desktop focused app and its focused state. Added a CheckBox "IntgrScanFsorc_checkbox" for showing the VR focused app and two TextBoxes "WindowActivityVRTitle" and "WindowActivityVRFocusTitle" for displaying the title of the VR focused app and its focused state. Adjusted the width of the "PrivateAppCustom" TextBox from a fixed value to auto. Removed Border and ItemsControl elements used for displaying a list of scanned apps.