Updated IntelliChatModule and version in MagicChatbox
Updated the `GetDefaultWritingStyles` method in `IntelliChatModule.cs` to include two new properties: `StyleDescription` and `IsBuiltIn`. Adjusted `Temperature` values for some styles. Also, updated a system message to specify that the text length should remain the same when rewriting a sentence in a specific style. Finally, the version number of the `MagicChatbox.csproj` project has been updated from `0.8.800` to `0.8.805`.