Initialized properties in PulsoidModule and PulsoidTrendSymbolSet
In this commit, the `selectedPulsoidTrendSymbol` property in the `PulsoidModule` class has been initialized with a new instance of `PulsoidTrendSymbolSet`. Additionally, default values have been set for the `UpwardTrendSymbol` and `DownwardTrendSymbol` properties in the `PulsoidTrendSymbolSet` class. The `UpwardTrendSymbol` is now initialized to "⤴️" and the `DownwardTrendSymbol` is initialized to "⤵️". A minor formatting change was also made to the closing bracket of the `PulsoidModule` class.