Added new exception logging and refactored event handlers
Imported the `System.Runtime.ExceptionServices` namespace. The event handlers for `AppDomain.CurrentDomain.UnhandledException` and `DispatcherUnhandledException` were removed and re-added, possibly due to refactoring or reordering of the code. A new event handler `CurrentDomain_FirstChanceException` was added to the `AppDomain.CurrentDomain` object, which logs the exception message and stack trace whenever an exception is thrown for the first time in the application domain. A new method with the same name was also added for this purpose.