"Update backupPath assignment and downgrade app version"
Modified the assignment of `backupPath` in `UpdateApp.cs` within the `vrcosc_magicchatbox.Classes.DataAndSecurity` namespace. Instead of using the `backupPath` key from the `appLocation` JSON object, it now uses the `Path.Combine` method to combine `dataPath` with the string "backup". Also, downgraded the application version in `MagicChatbox.csproj` from 0.8.825 to 0.8.819.