Added color customization feature to UI
A new namespace `vrcosc_magicchatbox.UI.Windows` was added along with a `ColorCustomizationWindow` in `App.xaml.cs`. A `ColorToBrushConverter` class was implemented in `ColorToBrushConverter.cs` to convert `Color` objects to `SolidColorBrush` and vice versa. The `Extended.Wpf.Toolkit` package version `4.6.0` was added to the project. `ColorCustomizationWindow.xaml.cs` file was added to the project. A `ColorToBrushConverter` resource was added to `MainWindow.xaml` and several UI elements had their color properties bound to properties in the `ColorCustomizationViewModel`. A `ColorCustomizationViewModel` class was added for color customization of the application's UI. A `ColorCustomizationWindow` was added for user interaction. The `ColorCustomizationWindow.xaml.cs` file was added for the code-behind. A `ColorCustomizationViewModel` property was added to the `ViewModel` class.