Added MediaLinkTimeSeekbar enum and updated UI
This commit introduces the `MediaLinkTimeSeekbar` enum in `MediaLinkModule.cs` to allow different styles for the media link time seek bar. The `MediaLinkShowTime` property in `ViewModel.cs` was replaced with `MediaLinkTimeSeekStyle` to reflect this change. A new property `AvailableTimeSeekbarStyles` was also added to `ViewModel.cs` to get all possible values of the new enum. The dictionary key in `DataController.cs` and the condition in `OSCController.cs` were updated accordingly. Two new methods `CreateTimeStamp()` and `CreateProgressBar()` were added to `OSCController.cs` to create a timestamp and a progress bar based on the `MediaLinkTimeSeekStyle` property. The UI in `MainWindow.xaml` was updated to replace the checkbox with a combo box for `MediaLinkTimeSeekStyle`, and minor adjustments were made to the width and margin of various UI elements.