Added system culture setting for time formatting
A new feature has been introduced that allows the use of the system's culture settings for time formatting. This was achieved by adding a new key-value pair to the dictionary in `DataController.cs` and a new `UseSystemCulture` property in `ViewModel.cs`. The `GetFormattedTime` method in `ComponentStatsModule.cs` was updated to use this new property, and its access modifier was changed from private to public. A CheckBox UI element was also added to `MainWindow.xaml` to allow users to toggle this setting.