Update time separator in smallNumbersContent string
The commit updates the formatting of the `smallNumbersContent` string in the `OSCController.cs` file, under the `vrcosc_magicchatbox.Classes.DataAndSecurity` namespace. The change involves replacing the colon (":") that separated `currentTime` and `fullTime` values with a vertical bar ("l"). This alteration affects the visual representation of the time span in the `smallNumbersContent` string, but does not impact the functionality of the code.